Update the sample app
